+++ src/test/disruptive.test.ts
@@ -255,7 +255,7 @@ async function createObj(
 }
 
 test("Create CDS objects", async () => {
-  if (!enableWrite(new Date())) fail("dummy")
+  if (!enableWrite(new Date())) return
   const c = create()
   const options: NewObjectOptions = {
     description: "test CDS AC creation",